+++ deployments/original-data-to-bigqueryload/tests/libs/gcp_pipeline_tester/bdd/__init__.py
@@ -1,7 +1,0 @@
-"""
-BDD Testing Utilities for GCP pipelines.
-"""
-
-from .base import PipelineScenarioTest
-
-__all__ = ["PipelineScenarioTest"]