+++ gcp-pipeline-libraries/gcp-pipeline-beam/src/gcp_pipeline_beam/__init__.py
@@ -1,4 +1,0 @@
-"""
-GCP Pipeline Beam Library
-"""
-__version__ = "1.0.29"